# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  auth.AppleLoginInput:
    properties:
      fullName:
        properties:
          familyName:
            type: string
          givenName:
            type: string
        type: object
      identityToken:
        type: string
    required:
    - identityToken
    type: object
  auth.LoginInput:
    description: User login request body
    properties:
      email:
        description: Email
        example: john@example.com
        type: string
      password:
        description: Password
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.LogoutInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshTokenInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.SetPasswordInput:
    properties:
      password:
        minLength: 8
        type: string
    required:
    - password
    type: object
  auth.loginSuccessData:
    properties:
      access_token:
        description: Access token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        description: Refresh token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  auth.registerSuccessData:
    properties:
      user_id:
        description: User ID
        example: 1
        type: integer
      username:
        description: Username
        example: john_doe
        type: string
    type: object
  auth.uploadAvatarSuccessData:
    properties:
      avatar_url:
        type: string
    type: object
...
  internal_api_v1_auth.RegisterInput:
    description: User registration request body
    properties:
      email:
        description: Email
        example: john@example.com
        type: string
      password:
        description: Password (min 8 characters)
        example: password123
        minLength: 8
        type: string
      username:
        description: Username
        example: john_doe
        type: string
    required:
    - email
    - password
    - username
    type: object
  response.ApiResponse:
    properties:
      code:
        description: 0 success, non-zero failure
        type: integer
      data:
        description: Optional payload
      message:
        description: Human-readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: PrismBox backend HTTP API reference
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PrismBox Backend API
  version: "1.0"
paths:
...
  /auth/apple/login:
    post:
      consumes:
      - application/json
      description: Returns access and refresh tokens using Apple identity token
      parameters:
      - description: Apple Sign In payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.AppleLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.loginSuccessData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Invalid Apple token
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Apple Sign In
      tags:
      - Auth
  /auth/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Uploads avatar image (authentication required); jpg, jpeg, or png,
        max 5MB
      parameters:
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.uploadAvatarSuccessData'
              type: object
        "400":
          description: Unsupported format or file too large
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Upload avatar
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns access and refresh tokens
      parameters:
      - description: Login payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.loginSuccessData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the provided refresh token
      parameters:
      - description: Logout payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Logout
      tags:
      - Auth
  /auth/password/set:
    post:
      consumes:
      - application/json
      description: Sets password for the authenticated user account
      parameters:
      - description: Password payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.SetPasswordInput'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request or password already set
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Set password
      tags:
      - Auth
  /auth/profile:
    get:
      description: Returns profile for the current user (authentication required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access token using a refresh token
      parameters:
      - description: Refresh token payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid or revoked refresh token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Refresh token expired
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new account with username, email, and password
      parameters:
      - description: Registration payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_api_v1_auth.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.registerSuccessData'
              type: object
        "400":
          description: Bad request or username/email already exists
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Register user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Use `Authorization: Bearer {token}` or header `x-prismbox-user-token:
      {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"